requirements.txt
fastmcp==2.13.0.2
openai==2.7.2
python-dotenv==1.2.1
streamlit == 1.51.0
